Add Jimmy Administrator whitelist
I've hard-coded this as a list of udernames, initially, for the sake of simplicity. Eventually, this may need to be migrated to a column of the user table. The name 'jimmy_admin' was chosen so as to avoid conflicts with potential other administrators int he future. This represents the creator/maintainer of the app, not customers who are administrators