Add Admin Flag to User
This adds an `admin?` flag tot he user. I've hard-coded this as a list
of udernames, initially, for the sake of simplicity. Eventually, this
may need to be migrated to a column of the user table.

This, along with a new `route_constraints` pattern is used to lock the
Sidekiq and PGHero panesl to admins only.

See https://guides.rubyonrails.org/routing.html#advanced-constraints for
more on constraints.

The admin constrain is directly required in the routes.rb in order to
discourage overuse of constrains like this. In general, they should only
be used when mounting stuff from engines.